added base for testing the validator
