+++ tdrs-backend/tdpservice/parsers/aggregates.py
@@ -1,8 +1,9 @@
 """Aggregate methods for the parsers."""
 from .row_schema import SchemaManager
 from .models import ParserError
-from .util import month_to_int, get_program_models, get_text_from_df, \
+from .util import month_to_int, \
     transform_to_months, fiscal_to_calendar, get_prog_from_section
+from .schema_defs.utils import get_program_models, get_text_from_df
 
 
 def case_aggregates_by_month(df, dfs_status):
